// src/orders/order.dto.ts
import { OrderType } from '@prisma/client';

export class InitTransactionDto {
  asset_id: string;
  wallet_id: string;
  shares: number;
  type: OrderType;
  price: number;
}

export class InputExecuteTransactionDto {
  order_id: string;
  status: 'OPEN' | 'CLOSED';
  related_investor_id: string;
  broker_investor_id: string;
  negotiated_shares: number;
  price: number;
}